requirements.txt
snowflake-connector-python==3.10.1
python-dotenv==1.0.1
pandas==2.2.3
pypdf==5.1.0
streamlit>=1.36.0
pyyaml>=6.0
openai>=1.0.0

